move check_qro to utils.c
and removed locator2longlat.[ch]
