+++ src/Makefile.am
@@ -20,7 +20,7 @@ tlf_SOURCES = \
 	initial_exchange.c \
 	keyer.c \
 	lancode.c last10.c listmessages.c log_to_disk.c log_utils.c \
-	logit.c logview.c locator2longlat.c \
+	logit.c logview.c \
 	main.c makelogline.c messagechange.c muf.c \
 	nicebox.c note.c netkeyer.c\
 	paccdx.c parse_logcfg.c prevqso.c printcall.c \
@@ -57,7 +57,7 @@ noinst_HEADERS = \
 	ignore_unused.h initial_exchange.h \
 	keyer.h keystroke_names.h \
 	lancode.h last10.h listmessages.h log_utils.h \
-	log_to_disk.h logit.h logview.h locator2longlat.h \
+	log_to_disk.h logit.h logview.h \
 	makelogline.h messagechange.h muf.h \
 	nicebox.h note.h netkeyer.h\
 	paccdx.h parse_logcfg.h prevqso.h printcall.h \
